feat(flash): allow the use of previewers other than bat
This commit adds a way of changing the FZF previewer by using the `-p`
flag. An example of this might look like:

$ flash -p cat

This will use cat has the FZF previewer instead of bat. However, it keeps bat as
the default program in all cases unless specifically changed by the `-p`
parameter. There are also checks to see if the provided binary actually
exists on the system. If it exists on the system, it checks if it
supported by flash. If it isn't supported by flash, it falls back onto
bat as the default previewer. For now, only bat and cat are supported but it.
can easily be expanded.